support the butter algorithm only, now all butter iir filters in biquad setting seem to be working. I will probably have to print a warning against using filters that aren't biquad due to numerical instability for traditional ba iir filters

+++ impl/fixed_point.h
@@ -40,9 +40,9 @@ typedef double     filter_coeff_t;
 typedef float     filter_data_t;
 typedef double     filter_accum_t;
 #else
-typedef long   filter_coeff_t;
-typedef int   filter_data_t;
-typedef long   filter_accum_t;
+typedef long long  filter_coeff_t;
+typedef long   filter_data_t;
+typedef long long  filter_accum_t;
 #endif /* ENABLE_FLOATING_POINT_MATH */
 
 #if ENABLE_FLOATING_POINT_MATH
@@ -50,7 +50,7 @@ typedef long   filter_accum_t;
 #define FROM_FIXED_POINT(x) (x)
 #else
 #define FIXED_POINT_FRACTIONAL_BITS ((sizeof(filter_data_t) * 8) - 1)
-#define FIXED_POINT_SCALING_FACTOR  (1UL << FIXED_POINT_FRACTIONAL_BITS)
+#define FIXED_POINT_SCALING_FACTOR  (1ULL << FIXED_POINT_FRACTIONAL_BITS)
 #define TO_FIXED_POINT(x)   ((x) * FIXED_POINT_SCALING_FACTOR)
 #define FROM_FIXED_POINT(x) ((x) >> FIXED_POINT_FRACTIONAL_BITS)
 #endif /* ENABLE_FLOATING_POINT_MATH */
